Move page models
Move page models one step up in the namespace hierarchy.
