+++ tests/LondonTravel.Site.Tests/Pages/HomePage.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Martin Costello, 2017. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-namespace MartinCostello.LondonTravel.Site.Integration.Pages
+namespace MartinCostello.LondonTravel.Site.Pages
 {
     using System.Collections.Generic;
     using System.Linq;
